+++ tracker/src/plausible.js
@@ -5,12 +5,15 @@ import { init as initAutocapture } from './autocapture'
 import { track } from './track'
 
 function init(overrides) {
+  initConfig(overrides || {})
+
   if (COMPILE_PLAUSIBLE_WEB && window.plausible && window.plausible.l) {
-    console.warn('Plausible analytics script was already initialized, skipping init')
+    if (config.logging) {
+      console.warn('Plausible analytics script was already initialized, skipping init')
+    }
     return
   }
 
-  initConfig(overrides)
   initEngagementTracking()
 
   if (!COMPILE_MANUAL || (COMPILE_CONFIG && config.autoCapturePageviews)) {